feat: restructure flagd to flagd & kube-flagd-proxy
