+++ cmd/doc/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"log"
 
-	"github.com/open-feature/flagd/cmd"
+	"github.com/open-feature/flagd/cmd/flagd/cmd"
 )
 
 const docPath = "./docs/configuration"
