+++ cmd/flagd/cmd/doc.go
